Fixed Asteroid generation bug, that caused users without the plugin being disconnected from servers.
The asteroid generator was changed from the own implemented one taken from keen, to the one keen already uses itself. Previous it was not possible due to restrictions to the classes, which caused the Shape provider being internal only. The plugin now uses C# reflection to excecute the methods and thus using Space Engineers implementation directly instead of indirectly, which doesnt require a new storage provider, that is not available to vanilla players. Since this was the case they got kicked.